requirements.txt
accelerate==0.24.1
appdirs==1.4.4
apricot-select==0.6.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
et-xmlfile==1.1.0
filelock==3.13.1
flair @ git+https://github.com/jusiro/FLAIR.git@e81d1d007756a3d90718577d32590826d30004d9
fonttools==4.46.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.6.12
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
nose==1.3.7
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==1.5.3
peft==0.6.2
Pillow==9.4.0
protobuf==4.25.1
pyparsing==3.1.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.13.2
segment-anything==1.0
sentry-sdk==1.37.1
setproctitle==1.3.3
smmap==5.0.1
sympy==1.12
threadpoolctl==3.2.0
timm==0.9.12
tokenizers==0.13.3
torch==1.13.1
torchaudio==0.13.1
torchcam==0.3.2
torchvision==0.14.1
tqdm==4.64.1
transformers==4.27.4
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.1.0
wandb==0.16.0
